move thumbQueue to dedicated file
